Support arbitrary columns in read_csv
Instead of fixing read_csv to the column names md, inc, and azi, default
to them instead, and allow just-in-time specification.

This lifts the column order restriction, and should handle more files.
Usually, it's easier to change small programs, e.g. in notebooks, than
it is to change data files.